refactor(devtools): share generated surface writer
