+++ devtools/render_agents.py
@@ -9,6 +9,7 @@ import sys
 from pathlib import Path
 
 from devtools.command_catalog import control_plane_command
+from devtools.render_support import write_if_changed
 
 INCLUDE_RE = re.compile(r"^\s*@(.+?)\s*$")
 ESCAPED_CHAR_RE = re.compile(r"\\(.)")
@@ -81,19 +82,6 @@ def render_document(*, input_path: Path, output_path: Path | None) -> str:
         f"Edit {display_input} and included files instead. -->\n\n"
     )
     return header + body
-
-
-def write_if_changed(output_path: Path, content: str) -> None:
-    try:
-        current = output_path.read_text(encoding="utf-8")
-    except FileNotFoundError:
-        current = None
-    if current == content:
-        return
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-    tmp_path = output_path.with_suffix(output_path.suffix + ".tmp")
-    tmp_path.write_text(content, encoding="utf-8")
-    tmp_path.replace(output_path)
 
 
 def main(argv: list[str] | None = None) -> int:
