+++ devtools/render_devtools_reference.py
@@ -12,6 +12,7 @@ from devtools.command_catalog import (
     featured_command_specs,
     grouped_command_specs,
 )
+from devtools.render_support import write_if_changed
 
 MARKER = "devtools-command-catalog"
 
@@ -79,19 +80,6 @@ def replace_marked_section(text: str, replacement: str) -> str:
         raise ValueError(f"marker block not found: {MARKER}")
     finish += len(end)
     return text[:start] + replacement + text[finish:]
-
-
-def write_if_changed(output_path: Path, content: str) -> None:
-    try:
-        current = output_path.read_text(encoding="utf-8")
-    except FileNotFoundError:
-        current = None
-    if current == content:
-        return
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-    tmp_path = output_path.with_suffix(output_path.suffix + ".tmp")
-    tmp_path.write_text(content, encoding="utf-8")
-    tmp_path.replace(output_path)
 
 
 def main(argv: list[str] | None = None) -> int:
