+++ devtools/render_docs_surface.py
@@ -15,6 +15,7 @@ from devtools.docs_surface import (
     REPO_GUIDE_ENTRIES,
     DocsEntry,
 )
+from devtools.render_support import write_if_changed
 
 README_MARKER = "docs-surface"
 GENERATED_NOTE = (
@@ -103,19 +104,6 @@ def replace_marked_section(text: str, *, marker: str, replacement: str) -> str:
         raise ValueError(f"marker block not found: {marker}")
     finish += len(end)
     return text[:start] + replacement + text[finish:]
-
-
-def write_if_changed(output_path: Path, content: str) -> None:
-    try:
-        current = output_path.read_text(encoding="utf-8")
-    except FileNotFoundError:
-        current = None
-    if current == content:
-        return
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-    tmp_path = output_path.with_suffix(output_path.suffix + ".tmp")
-    tmp_path.write_text(content, encoding="utf-8")
-    tmp_path.replace(output_path)
 
 
 def render_outputs(*, readme_path: Path, docs_readme_path: Path) -> tuple[str, str]:
